requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.10.3
python-dotenv==1.0.1
supabase==2.10.0
openai==1.55.3
python-docx==1.1.2
